Feat/order

+++ src/main/java/com/scanly/scanlyBackend/dtos/ProductResponse.java
@@ -1,7 +1,9 @@
 package com.scanly.scanlyBackend.dtos;
 
+import java.math.BigDecimal;
+
 public record ProductResponse(
         String barcode,
         String name,
-        Double price
+        BigDecimal price
 ) {}
